Rename neon to neon-serverless, fix release pipeline
