+++ .github/workflows/release-latest.yaml
@@ -4,7 +4,9 @@ on: workflow_dispatch
 
 jobs:
   release:
+    permissions: write-all
     strategy:
+      fail-fast: false
       matrix:
         package:
           - drizzle-orm
@@ -23,7 +25,10 @@ jobs:
           version: latest
           run_install: true
 
-      - name: Publish
+      - name: Check preconditions
+        id: checks
+        shell: bash
+        working-directory: ${{ matrix.package }}
         run: |
           latest="$(npm view --json ${{ matrix.package }} dist-tags.latest | jq -r)"
           version="$(jq -r .version package.json)"
@@ -32,33 +37,49 @@ jobs:
             echo "Latest:  $latest"
             echo "Current: $version"
 
-            is_version_published="$(npm view ${{ matrix.package }} versions --json | jq -r '.[] | select(. == "'$version'") | . == "'$version'"')"
-            if [[ "$is_version_published" == "true" ]]; then
-              echo "Version $version already published, adding tag $tag"
-              npm dist-tag add ${{ matrix.package }}@$version latest
-            else
-              echo echo "Publishing ${{ matrix.package }}@$version"
-              (cd .. && npm run pack -- --filter ${{ matrix.package }})
-              npm run publish
+            changelogPath=$(node -e "console.log(require('path').resolve('..', 'changelogs', '${{ matrix.package }}', '$version.md'))")
+            if [[ ! -f "$changelogPath" ]]; then
+              echo "::error::Changelog for version $version not found: $changelogPath"
+              exit 1
             fi
 
-            # Post release message to Discord
-            curl -X POST -H "Content-Type: application/json" -d "{\"embeds\": [{\"title\": \"New \`${{ matrix.package }}\` release! 🎉\", \"url\": \"https://www.npmjs.com/package/${{ matrix.package }}\", \"color\": \"12907856\", \"fields\": [{\"name\": \"Tag\", \"value\": \"\`$tag\`\"}]}]}" ${{ secrets.DISCORD_RELEASE_WEBHOOK_URL }}
-
-            # Put version to outputs
-            echo "::set-output name=version::$version"
-            echo "::set-output name=has_new_release::true"
+            {
+              echo "version=$version"
+              echo "has_new_release=true"
+              echo "changelog_path=$changelogPath"
+            } >> $GITHUB_OUTPUT
           else
             echo "Already up to date: $version"
+            echo "\`$version\` is already latest on NPM" >> $GITHUB_STEP_SUMMARY
           fi
+
+      - name: Publish
+        if: steps.checks.outputs.has_new_release == 'true'
         working-directory: ${{ matrix.package }}
         shell: bash
         env:
           NODE_AUTH_TOKEN: ${{ secrets.NPM_ACCESS_TOKEN }}
+        run: |
+          version="${{ steps.checks.outputs.version }}"
+          is_version_published="$(npm view ${{ matrix.package }} versions --json | jq -r '.[] | select(. == "'$version'") | . == "'$version'"')"
 
-      - name: Create GitHub release
+          if [[ "$is_version_published" == "true" ]]; then
+            echo "Version $version already published, adding tag $tag"
+            npm dist-tag add ${{ matrix.package }}@$version latest
+          else
+            echo echo "Publishing ${{ matrix.package }}@$version"
+            (cd .. && npm run pack -- --filter ${{ matrix.package }})
+            npm run publish
+          fi
+
+          echo "npm: \`+ ${{ matrix.package }}@$version\`" >> $GITHUB_STEP_SUMMARY
+
+          # Post release message to Discord
+          curl -X POST -H "Content-Type: application/json" -d "{\"embeds\": [{\"title\": \"New \`${{ matrix.package }}\` release! 🎉\", \"url\": \"https://www.npmjs.com/package/${{ matrix.package }}\", \"color\": \"12907856\", \"fields\": [{\"name\": \"Tag\", \"value\": \"\`$tag\`\"}]}]}" ${{ secrets.DISCORD_RELEASE_WEBHOOK_URL }}
+
+      - name: Create GitHub release for ORM package
         uses: actions/github-script@v6
-        if: ${{ steps.release.outputs.has_new_release == 'true' }}
+        if: matrix.package == 'drizzle-orm' && steps.checks.outputs.has_new_release == 'true'
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           script: |
@@ -66,24 +87,24 @@ jobs:
               const fs = require("fs");
               const path = require("path");
 
-              const changelog = fs.readFileSync(path.join(__dirname, "changelogs", "${{ matrix.package }}", "${{ steps.release.outputs.version }}.md"), "utf8");
-              const version = "${{ steps.release.outputs.version }}";
-              const tag = "latest";
+              const version = "${{ steps.checks.outputs.version }}";
+              const changelog = fs.readFileSync("${{ steps.checks.outputs.changelog_path }}", "utf8");
+
               const release = await github.rest.repos.createRelease({
                 owner: context.repo.owner,
                 repo: context.repo.repo,
-                tag_name: tag,
-                name: `${{ matrix.package }} v${version}`,
+                tag_name: `${version}`,
+                name: `${version}`,
                 body: changelog,
-                draft: true,
               });
+
               await github.rest.repos.uploadReleaseAsset({
                 owner: context.repo.owner,
                 repo: context.repo.repo,
                 release_id: release.data.id,
-                name: `${{ matrix.package }}-${version}.tgz`,
-                data: fs.readFileSync(path.join(__dirname, "${{ matrix.package }}", "package.tgz")),
+                name: `${{ matrix.package }}-${version}-dist.tgz`,
+                data: fs.readFileSync(path.resolve("${{ matrix.package }}", "package.tgz")),
               });
             } catch (e) {
-              core.setFailed(error.message);
+              core.setFailed(e.message);
             }
